server/requirements.txt
flask
telethon
pillow
